feat:api경로 라우팅 기능 추가
기존 api 대부분 라우트파일로 이동,
post list 응답값 변경
