+++ Express 2/src/app.ts
@@ -1,6 +1,8 @@
 import express, { Application, Request, Response, NextFunction } from 'express';
 import path from 'path';
-import { send } from 'process';
+import 'tsconfig-paths/register';
+import userRouter from './routes/user';
+import postRouter from './routes/post'
 import swaggerUi from 'swagger-ui-express';
 import YAML from 'yamljs';
 
@@ -21,23 +23,30 @@ app.use(methodOverride('_method'));
 app.use(flash());
 app.set('view engine', 'ejs');
 
-let db: any;
+
+let connectDB: any;
 MongoClient.connect(
     process.env.DB_URL,
     { useUnifiedTopology: true },
     (err: Error, client: any) => {
         // 에러날 시
         if (err) return console.log(err);
-
+        
         // hansei db에 연결하기
-        db = client.db('hansei');
+            connectDB = client.db('hansei');
 
-        // 서버 실행
-        app.listen(process.env.PORT, () => {
-            console.log(`🛡️  Server listening on port: 8000🛡️`);
-        });
+        
     }
-);
+)
+export const db = connectDB
+app.set('db', db)
+
+app.use('/user', userRouter)
+app.use('/post', postRouter)
+// 서버 실행
+app.listen(process.env.PORT, () => {
+    console.log(`🛡️  Server listening on port: 8000🛡️`);
+});
 
 // 메인 페이지
 app.get('/', (req: Request, res: Response, next: NextFunction) => {
@@ -45,9 +54,9 @@ app.get('/', (req: Request, res: Response, next: NextFunction) => {
 });
 
 // 회원 인증
-const passport = require('passport');
+export const passport = require('passport');
 const LocalStrategy = require('passport-local').Strategy;
-const session = require('express-session');
+export const session = require('express-session');
 const FileStore = require('session-file-store')(session);
 const cors = require('cors');
 
@@ -139,73 +148,6 @@ passport.deserializeUser((id: any, done: any) => {
 //         res.status(401).json({ message: '로그인 실패' });
 //     }
 // });
-app.post(
-    '/user/login',
-    passport.authenticate('local'),
-    (req: Request | any, res: Response, next: NextFunction) => {
-        const { id, pw, studentId, department } = req.body; // 요청에서 아이디와 비밀번호 추출
-        console.log(id, pw, studentId, department, '요청받음');
-
-        req.login(req.user, (err: Error) => {
-            if (err) {
-                return next(err);
-            }
-        });
-        console.log('요청출력', id, pw, studentId, department);
-        req.login(req.user, (err: Error) => {
-            if (err) {
-                // res.setHeader('Content-Type', 'application/json');
-                // res.status(401).json({ data: '로그인실패' });
-                return next(err);
-            }
-        });
-        try {
-            res.header('Access-Control-Allow-Credentials', 'true');
-            res.status(200).json({ id: id, message: '정상로그인' });
-            console.log('로그인성공');
-        } catch (error) {
-            console.log('로그인 오류');
-        }
-    }
-);
-
-// /login - 로그인 페이지
-app.get(
-    '/user/login',
-    loginUser,
-    (req: Request | any, res: Response, next: NextFunction) => {
-        if (req.headers.origin === 'localhost:8000') {
-            var user = req.session.user;
-            // 세션 데이터 활용
-            if (user) {
-                // 로그인된 사용자에 대한 프로필 페이지 보여주기
-                res.render('mypage.ejs', { user });
-            } else {
-                // 로그인되지 않은 사용자는 로그인 페이지로 리다이렉트
-                // res.redirect('/user/login');
-            }
-            res.render('login.ejs');
-        }
-        console.log('인증성공');
-        res.status(200).json({ succeed: true, message: '로그인된 유저입니다' });
-    }
-);
-
-app.delete(
-    '/user/login',
-    loginUser,
-    (req: Request | any, res: Response, next: NextFunction) => {
-        req.session.destroy((err: any) => {
-            if (err) {
-                // 세션 삭제 실패
-                console.log(err);
-            }
-            // 세션 삭제 성공
-            console.log('세션 삭제 완료');
-            res.status(204).end();
-        });
-    }
-);
 
 app.get(
     '/user/fail',
@@ -219,299 +161,28 @@ app.get(
     }
 );
 
-// /mypage - 마이 페이지
-app.get(
-    '/user/mypage',
-    loginUser,
-    (req: Request | any, res: Response, next: NextFunction) => {
-        console.log('로그인된 유저');
-        res.header('Content-Type', 'application/json');
-        res.status(200).json({ succeed: true, user: { id: req.user.id } });
-
-        if (req.headers.origin == 'localhost:8000') {
-            res.render('mypage.ejs', { userMe: req.user });
-        }
-    }
-);
-// 마이페이지 접속전 미들웨어
-function loginUser(req: Request | any, res: Response, next: any) {
-    if (req.user) {
-        // 요청 user 있으면 통과
-        next();
-    } else {
-        res.status(401).json({ succeed: false, message: '로그인 안함!' });
-    }
-}
-
-// /join - 회원가입 페이지
-app.get(
-    '/user/join',
-    (req: Request | any, res: Response, next: NextFunction) => {
-        res.render('join.ejs');
-    }
-);
-
-app.post('/user/join', (req: Request, res: Response, next: NextFunction) => {
-    // login 컬렉션에 회원가입 정보 저장하기
-    db.collection('login').insertOne(
-        {
-            id: req.body.id,
-            pw: req.body.pw,
-            studentId: req.body.studentId,
-            department: req.body.department,
-        },
-        (err: Error, result: any) => {
-            console.log('저장완료!!!');
-        }
-    );
-    res.redirect('/login');
-});
 
 // /search - 검색 페이지
-app.get('/post/search', (req: Request, res: Response) => {
-    // 검색어에 숫자가 있어도 일치시키기 위한 정규식
-    const searchValue: any = req.query.value; // 타입을 명시적으로 지정
-    console.log(typeof searchValue);
-    if (!searchValue) {
-        return res.status(400).send('검색어를 입력해주세요.');
-    }
 
-    const regexPattern = new RegExp(searchValue, 'i');
-    console.log(regexPattern);
-
-    let searchPost = [
-        {
-            $search: {
-                index: 'titleSearch',
-                text: {
-                    query: searchValue,
-                    path: '제목', // 제목날짜 둘다 찾고 싶으면 ['제목', '날짜']
-                },
-            },
-        },
-        { $sort: { _id: 1 } },
-        { $limit: 10 },
-    ];
-    // console.log(req.query.value);
-    db.collection('post')
-        .aggregate(searchPost)
-        .toArray((err: Error, result: any) => {
-            if (err) {
-                // 오류 처리
-                console.log(err);
-
-                return res
-                    .status(500)
-                    .send('검색 결과를 가져오는 도중 오류가 발생했습니다.');
-            }
-            console.log('검색 결과:', result);
-            console.log('검색 정규식:', regexPattern);
-            res.render('search.ejs', { posts: result ?? [] });
-        });
-});
 
 // /write 글 작성 페이지
 app.get('/post/write', (req: Request, res: Response, next: NextFunction) => {
-    res.render('write.ejs');
+    if(req.headers.origin=='localhost:8000')res.render('write.ejs');
 });
 
 // /write - form 데이터 /write-page로 POST 요청
-app.post(
-    '/post/write-page',
-    (req: Request | any, res: Response, next: NextFunction) => {
-        db.collection('count').findOne(
-            { name: '게시물갯수' },
-            (err: Error, result: any) => {
-                let total = result.totalPost;
-                let storage = {
-                    _id: total + 1,
-                    // 현재 로그인한 사람 정보
-                    작성자: req.user.id,
-                    제목: req.body.title,
-                    날짜: req.body.date,
-                    내용: req.body.content,
-                };
-                db.collection('post').insertOne(
-                    storage,
-                    (err: any, result: any) => {
-                        db.collection('count').updateOne(
-                            { name: '게시물갯수' },
-                            { $inc: { totalPost: 1 } },
-                            (err: Error, result: any) => {
-                                if (err) {
-                                    return console.log(err);
-                                }
-                                res.redirect('/list');
-                            }
-                        );
-                    }
-                );
-            }
-        );
-    }
-);
+
 
 // /list GET 요청 처리
-app.get('/post/list', (req: Request, res: Response, next: NextFunction) => {
-    // 작성자 정보 가져오기
-    db.collection('login')
-        .find()
-        .toArray((err: Error, loginResult: any) => {
-            if (err) {
-                // 오류 처리
-                return next(err);
-            }
-
-            // 게시물 가져오기
-            db.collection('post')
-                .find()
-                .toArray((err: Error, postResult: any) => {
-                    if (err) {
-                        // 오류 처리
-                        return next(err);
-                    }
-
-                    // 게시물 작성자 정보를 포함하여 렌더링
-                    res.render('list.ejs', {
-                        loginData: loginResult,
-                        posts: postResult,
-                        getAuthorName: getAuthorName,
-                    });
-                });
-        });
-});
 
 // getAuthorName 함수 정의
-function getAuthorName(authorId: any, loginData: any) {
-    // 작성자 id를 사용하여 작성자 이름을 가져오는 로직 구현
-    const author =
-        loginData && loginData.find((login: any) => login.id === authorId);
-    return author ? author.name : '';
-}
+
 
 // /delete
-app.delete('/post/delete', (req: Request | any, res: Response) => {
-    // db에서 삭제하기
-    console.log(req.body);
 
-    // 로그인 사용자 ID 확인
-    const loggedInUserId = req.user.id;
-    // 클라이언트에서 전달된 _id 값
-    const postId = parseInt(req.body._id);
-    // 실제 로그인 유저 ID와 글에 저장된 작성자 ID 일치 여부 확인
-    const deleteData = { _id: postId, 작성자: loggedInUserId };
-
-    console.log('작성자 ID:', loggedInUserId);
-    console.log('_id 값:', postId);
-
-    db.collection('post').findOne(deleteData, (err: Error, post: any) => {
-        if (err) {
-            console.log(err);
-            res.status(500).send({ message: '오류가 발생했습니다.' });
-            return;
-        }
-
-        if (!post) {
-            // 작성자와 일치하는 글이 없는 경우
-            res.status(403).send({ message: '내가 쓴 글이 아닙니다.' });
-            return;
-        }
-
-        // 일치하는 글이 있는 경우 삭제 수행
-        db.collection('post').deleteOne(
-            deleteData,
-            (err: Error, result: any) => {
-                if (err) {
-                    console.log(err);
-                    res.status(500).send({ message: '오류가 발생했습니다.' });
-                    return;
-                }
-
-                res.status(200).send({ message: '성공했습니다.' });
-            }
-        );
-    });
-});
 
 // /edit - 수정페이지
-app.get(
-    '/post/edit/:id',
-    (req: Request | any, res: Response, next: NextFunction) => {
-        const postId = parseInt(req.params.id);
-        const loggedInUserId = req.user?.id;
 
-        db.collection('post').findOne(
-            { _id: postId },
-            (err: Error, result: any) => {
-                if (err) {
-                    console.error(err);
-                    return res
-                        .status(500)
-                        .json({ message: '에러가 발생했습니다.' });
-                }
-                // 작성자와 로그인한 사용자가 같은 경우에만 수정 페이지를 렌더링
-
-                if (result && result.작성자 === loggedInUserId) {
-                    res.render('edit.ejs', { post: result });
-                } else {
-                    // 작성자와 로그인한 사용자가 다른 경우, 또는 글이 없는 경우 에러 메시지를 전송
-                    res.status(403).send({
-                        message: '수정할 수 있는 권한이 없습니다.',
-                    });
-                }
-            }
-        );
-    }
-);
-app.put(
-    '/post/edit',
-    (req: Request | any, res: Response, next: NextFunction) => {
-        const postId = parseInt(req.body.id);
-        const loggedInUserId = req.user?.id;
-
-        db.collection('post').findOne(
-            { _id: postId },
-            (err: Error, result: any) => {
-                if (err) {
-                    return next(err);
-                }
-
-                // 작성자와 로그인한 사용자가 같은 경우에만 글을 수정합니다.
-                if (result && result.작성자 === loggedInUserId) {
-                    db.collection('post').updateOne(
-                        { _id: postId },
-                        {
-                            $set: {
-                                제목: req.body.title,
-                                날짜: req.body.date,
-                                내용: req.body.content,
-                            },
-                        },
-                        (err: Error, result: any) => {
-                            if (err) {
-                                return next(err);
-                            }
-                            console.log('수정!');
-                            // /list로 이동
-                            res.redirect('/list');
-                        }
-                    );
-                } else {
-                    res.status(403).send({
-                        message: '수정할 수 있는 권한이 없습니다.',
-                    });
-                }
-            }
-        );
-    }
-);
 
 // /detail - 상세 글 페이지
-app.get('/post/detail/:id', (req: Request, res: Response) => {
-    db.collection('post').findOne(
-        { _id: parseInt(req.params.id) },
-        (err: Error, result: any) => {
-            res.render('detail.ejs', { data: result });
-        }
-    );
-});
+
