+++ Express 2/views/write.ejs
@@ -17,7 +17,7 @@
         <h4 class="container mt-4"><strong>글 작성페이지</strong></h4>
 
         <div class="container mt-3">
-            <form action="/post/write-page" method="POST">
+            <form action="/post/write" method="POST">
                 <div class="form-group">
                     <label>오늘의 할일</label>
                     <input type="text" class="form-control" name="title" />
